requirements.txt
Flask
gunicorn
pandas
numpy
PyYaml
-e .